separate firebase logic in a new package
