+++ flutter_cache_manager/pubspec.yaml
@@ -18,7 +18,6 @@ dependencies:
   clock: ^1.0.1
   file: ^5.1.0
   rxdart: '>=0.23.1 <0.25.0'
-  firebase_storage: ^3.0.3
   
 dev_dependencies:
   flutter_test:
