+++ flutter_cache_manager_firebase/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 2738a1148ba6c9a6114df62358109407c3ef2553
+  channel: beta
+
+project_type: package
