+++ flutter_cache_manager_firebase/CHANGELOG.md
@@ -1,0 +1,3 @@
+## [1.0.0] - Initial release
+
+* Support for Firebase storage
