+++ flutter_cache_manager_firebase/lib/flutter_cache_manager_firebase.dart
@@ -1,0 +1,4 @@
+library flutter_cache_manager_firebase;
+
+export 'src/firebase_cache_manager.dart';
+export 'src/firebase_http_file_service.dart';
