+++ flutter_cache_manager_firebase/test/flutter_cache_manager_firebase_test.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:flutter_cache_manager_firebase/flutter_cache_manager_firebase.dart';
+
+void main() {
+}
